khai_dai/requirements.txt
asgiref
beautifulsoup4
certifi
crispy-bootstrap5
distlib
dj-database-url
Django
django-bootstrap4
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
EasyProcess
entrypoint2
filelock
Markdown
MouseInfo
mss
Pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
pyscreenshot
PyScreeze
python-decouple
pytweening
pytz
soupsieve
sqlparse
typing_extensions
tzdata
virtualenv==20.24.7
